+++ src/matrix/mod.rs
@@ -11,7 +11,7 @@ mod math;
 #[derive(Debug)]
 pub struct Matrix<T> {
     pub values: Vec<T>,
-    pub shape : (i32,i32),
+    pub shape : (usize,usize),
 }
